refractor new UploadButton component to handle all upload related tasks and codes
